requirements.txt
# Core dependencies for AI Agent
google-adk>=0.1.0
google-generativeai>=0.3.0
jupyterlab>=4.0.0

# Helper libraries
python-dotenv>=1.0.0
